requirements.txt
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.1
asttokens==2.0.8
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
cliapp==1.20180812.1
Click==7.0
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
contourpy==1.0.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==1.0.0
Django==3.2.15
django-crispy-forms==1.14.0
django-geoposition==0.3.0
django-heroku==0.3.1
django-leaflet==0.28.3
django-rest-knox==4.2.0
django-utils-six==2.0
djangorestframework==3.14.0
drf-yasg==1.21.4
duplicity==0.8.12.0
entrypoints==0.3
executing==1.1.1
fasteners==0.14.1
filelock==3.0.12
fonttools==4.38.0
future==0.18.2
GDAL==3.0.4
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
inflection==0.5.1
ipykernel==6.16.0
ipython==8.5.0
ipywidgets==8.0.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.4.2
jupyterlab-widgets==3.0.3
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
monotonic==1.5
more-itertools==4.2.0
nest-asyncio==1.5.6
netifaces==0.10.4
numpy==1.23.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
pandas==1.5.1
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
phonenumbers==8.12.47
pickleshare==0.7.5
Pillow==7.0.0
pipenv==11.9.0
prompt-toolkit==3.0.31
protobuf==3.6.1
psutil==5.9.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycairo==1.16.2
pycups==1.9.73
Pygments==2.13.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.5
pyxdg==0.26
PyYAML==5.3.1
pyzmq==24.0.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.1
systemd-python==234
tornado==6.2
traitlets==5.4.0
ttystatus==0.38
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
wcwidth==0.2.5
whitenoise==6.2.0
widgetsnbextension==4.0.3
xkit==0.0.0
zipp==1.0.0
